Centre the image over the document icon rather than placing it on top
Currently 'Latest Additions' is displayed with the document thumbnail
(at full width) and then a faded `ep_public_doc` icon is added below it
if there is space. This changes it so that the `ep_public_doc` icon is
always centered in the background and the thumbnail is shown on top of
it. It also zooms the thumbnail out if it is taller than the box (thats
what `object-fit: contain` does).