Allow the tile zoom on hover to be disabled in config
Adds a config option `latest_tool_tile_zoom` which then gets passed
through to the template and eventually HTML as the class
`ep_citation_tile_zoom` which is used by the CSS to control whether or
not it listens to `:hover` to zoom the images in.

It's potentially worth making this more general so that there is for
example an option to fill the box entirely or center the image, which
could be useful for repositories with a lot of images rather than a lot
of PDFs where this default isn't necessarily the best.

Fixes #13.