Override the core `latest_tool` cgi on 3.4
The `cgi/latest_tool` file is being removed from 3.5 core as it includes
this ingredient by default however we can't do that for 3.4 as it
doesn't automatically add this ingredient. However, due to e4ffae8
(cherry-picked from bb0261e) this will cause a citation error if you add
it to a server using `cgi/latest_tool` rather than `cgi/z_latest_tool`.

While we can't remove the equivalent file in core, we can rename this
one to override it, which as it is effectively identical (but sets the
necessary `tile_zoom` citation variable) shouldn't make any difference.

The only potential issue I can see is if someone is directly pointing at
the `z_latest_tool` cgi for which the solution would just be to update
this to point to `latest_tool` (which it could be pointing to already).
If this is a major issue for some reason we could just create a copy or
symlink at `z_latest_tool` pointing to `latest_tool`.