Don't zoom on hover with `prefers-reduced-motion` set
While eprints/latest_additions#15 allows the server to disable zoom on
hover for the tile view, it doesn't allow the client any way to override
the server's choice and disable this zooming.

There is a media query for this (`prefers-reduced-motion`) which is
widely supported so we should be checking that the user hasn't said they
don't want excessive movement before doing any zooming (as it is fully
unnecessary visual pizzazz). In the vein of it being fully unnecessary,
if the users browser does not support the `prefers-reduced-motion` query
(highly unlikely at this point, see
[caniuse](https://caniuse.com/?search=prefers-reduced-motion)) then
zooming on hover will simply be disabled.

This fixes this breakage of WCAG
[SC 2.3.3](https://www.w3.org/WAI/WCAG21/Understanding/animation-from-interactions.html#resources)
as per [Technique C39](https://www.w3.org/WAI/WCAG21/Techniques/css/C39).